trained linear regression model example

+++ src/auto_grad_node_impl/mean.cpp
@@ -5,7 +5,7 @@ namespace mango {
 MeanBackward::MeanBackward(Tensor a) { parents_.push_back(std::move(a)); }
 
 void MeanBackward::backwardPass(const Tensor &grad_out) {
-  Tensor meanDerivative({1.f / static_cast<float>(parents_[0].numel())},
+  Tensor meanDerivative({1.0 / static_cast<double>(parents_[0].numel())},
                         Shape{});
   parents_[0].accum_grad(grad_out.mult(meanDerivative));
   if (parents_[0].grad_fn()) {
